Fix AWS SSO provider
The AWS SSO provider was not working on Android, because:
- The browser window we opened to verify the device was not closeable.
  This is fixed, by using the default browser on Android instead of the
  in app browser.
- The "getSSOToken" method failed, because we used the wrong type in the
  Kotlin code. There we have to use kotlin.Long instead of kotlin.Int.

+++ android/app/src/main/kotlin/io/kubenav/kubenav/KubenavPlugin.kt
@@ -133,7 +133,7 @@ class KubenavPlugin : FlutterPlugin, MethodCallHandler {
       val ssoClientSecret = call.argument("ssoClientSecret") as String?
       val ssoDeviceCode = call.argument("ssoDeviceCode") as String?
       val accessToken = call.argument("accessToken") as String?
-      val accessTokenExpire = call.argument("accessTokenExpire") as kotlin.Int?
+      val accessTokenExpire = call.argument("accessTokenExpire") as kotlin.Long?
 
       if (accountID == null || roleName == null || ssoRegion == null || ssoClientID == null || ssoClientSecret == null || ssoDeviceCode == null || accessToken == null || accessTokenExpire == null) {
         result.error("BAD_ARGUMENTS", null, null)
@@ -276,7 +276,7 @@ class KubenavPlugin : FlutterPlugin, MethodCallHandler {
     }
   }
 
-  private fun awsGetSSOToken(accountID: String, roleName: String, ssoRegion: String, ssoClientID: String, ssoClientSecret: String, ssoDeviceCode: String, accessToken: String, accessTokenExpire: kotlin.Int, result: MethodChannel.Result) {
+  private fun awsGetSSOToken(accountID: String, roleName: String, ssoRegion: String, ssoClientID: String, ssoClientSecret: String, ssoDeviceCode: String, accessToken: String, accessTokenExpire: kotlin.Long, result: MethodChannel.Result) {
     try {
       val data: String = Kubenav.awsGetSSOToken(accountID, roleName, ssoRegion, ssoClientID, ssoClientSecret, ssoDeviceCode, accessToken, accessTokenExpire.toLong())
       result.success(data)
